[PRD-5886] View Menu - The Grid Settings dialog fails to open because of java.lang.ClassCastException: java.lang.Double cannot be cast to java.lang.Integer

+++ designer/report-designer/src/main/java/org/pentaho/reporting/designer/core/util/GridSizeDialog.java
@@ -12,7 +12,7 @@
 * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Lesser General Public License for more details.
 *
-* Copyright (c) 2002-2013 Pentaho Corporation..  All rights reserved.
+* Copyright (c) 2002-2017 Pentaho Corporation..  All rights reserved.
 */
 
 package org.pentaho.reporting.designer.core.util;
@@ -164,7 +164,7 @@ public class GridSizeDialog extends CommonDialog {
       throw new NullPointerException();
     }
     final double gridSize = getGridSize();
-    final double gridDivision = getGridDivisionSize();
+    final int gridDivision = getGridDivisionSize();
 
     if ( newUnit != this.unitModel.getSelectedItem() ) {
       this.unitModel.setSelectedItem( newUnit );
@@ -179,7 +179,7 @@ public class GridSizeDialog extends CommonDialog {
     this.gridSizeSpinner.setValue( new Double( unit.convertFromPoints( gridSize ) ) );
   }
 
-  public void setGridDivisionSize( final double gridDivisionSize ) {
+  public void setGridDivisionSize( final int gridDivisionSize ) {
     this.gridDivisionSpinner.setValue( gridDivisionSize );
   }
